build-logic 힐트 커스텀 플러그인 개선 (#23)
* #22 - feat : HiltAndroid 파일의 configureHilt를 모든 모듈에서 사용가능하도록 변경

* #22 - build : hilt 설정하는 플러그인의 이름 변경

* #22 - feat : 사용하지않은 Kotlin 모듈용 플러그인 삭제

* #22 - style : EOF 수정
