+++ build-logic/src/main/java/com/hegunhee/maplemfinder/build_logic/maplefinder.android.application.gradle.kts
@@ -1,5 +1,5 @@
 import com.hegunhee.maplemfinder.build_logic.setting.configureComposeAndroid
-import com.hegunhee.maplemfinder.build_logic.setting.configureHiltAndroid
+import com.hegunhee.maplemfinder.build_logic.setting.configureHilt
 import com.hegunhee.maplemfinder.build_logic.setting.configureKotlinAndroid
 
 plugins {
@@ -8,4 +8,4 @@ plugins {
 
 configureKotlinAndroid()
 configureComposeAndroid()
-configureHiltAndroid()
+configureHilt()
