+++ build-logic/src/main/java/com/hegunhee/maplemfinder/build_logic/maplefinder.android.library.gradle.kts
@@ -1,4 +1,4 @@
-import com.hegunhee.maplemfinder.build_logic.setting.configureHiltAndroid
+import com.hegunhee.maplemfinder.build_logic.setting.configureHilt
 import com.hegunhee.maplemfinder.build_logic.setting.configureKotlinAndroid
 
 plugins {
@@ -6,4 +6,4 @@ plugins {
 }
 
 configureKotlinAndroid()
-configureHiltAndroid()
+configureHilt()
