--- build-logic/src/main/java/com/hegunhee/maplemfinder/build_logic/setting/HiltAndroid.kt
@@ -1,28 +1,0 @@
-package com.hegunhee.maplemfinder.build_logic.setting
-
-import com.hegunhee.maplemfinder.build_logic.setup.libs
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.gradle.kotlin.dsl.dependencies
-
-internal fun Project.configureHiltAndroid() {
-    with(pluginManager) {
-        apply("dagger.hilt.android.plugin")
-        apply("org.jetbrains.kotlin.kapt")
-    }
-
-    val libs = extensions.libs
-    dependencies {
-        "implementation"(libs.findLibrary("hilt.android").get())
-        "kapt"(libs.findLibrary("hilt.compiler").get())
-    }
-}
-
-internal class HiltAndroidPlugin : Plugin<Project> {
-
-    override fun apply(target: Project) {
-        with(target) {
-            configureHiltAndroid()
-        }
-    }
-}
