+++ gradle/libs.versions.toml
@@ -24,7 +24,6 @@ ext-junit = "1.1.5"
 espresso-core = "3.5.1"
 
 hilt = "2.44.2"
-javax-inject = "1"
 
 okhttp = "4.11.0"
 retrofit = "2.9.0"
@@ -72,7 +71,6 @@ ui-test-manifest = {module = "androidx.compose.ui:ui-test-manifest"}
 hilt-core = { module = "com.google.dagger:hilt-core", version.ref = "hilt" }
 hilt-compiler = { module = "com.google.dagger:hilt-compiler", version.ref = "hilt" }
 hilt-android = { module = "com.google.dagger:hilt-android", version.ref = "hilt" }
-javax-inject = {module ="javax.inject:javax.inject",version.ref ="javax-inject"}
 
 okhttp-logging = { module = "com.squareup.okhttp3:logging-interceptor", version.ref = "okhttp" }
 retrofit2-retrofit = {module = "com.squareup.retrofit2:retrofit",version.ref ="retrofit"}
